UI Stuff ~UI Progress Bars (Health) ~UI Buttons ~UI Button Events ~UIButton Actions

+++ Rain/src/com/aadi/rain/entity/projectile/Projectile.java
@@ -7,7 +7,7 @@ import com.aadi.rain.graphics.Sprite;
 
 public abstract class Projectile extends Entity {
 
-	protected final double xOrigin, yOrigin;
+	protected final int xOrigin, yOrigin;
 	protected double angle;
 	protected Sprite sprite;
 	protected double x, y;
@@ -17,11 +17,10 @@ public abstract class Projectile extends Entity {
 
 	protected final Random random = new Random();
 
-	public Projectile(double x, double y, double dir) {
+	public Projectile(int x, int y, double dir) {
 		xOrigin = x;
 		yOrigin = y;
 		angle = dir;
-
 		this.x = x;
 		this.y = y;
 	}
